requirements.txt

torch==1.10.0
torchvision==0.11.1
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
matplotlib==3.4.3
joblib==1.0.1
scikit-image==0.18.2
scikit-learn==0.24.2
pandas==1.3.3
